Extract JsonCodecs to separate file and group api classes in package object
