+++ src/test/scala/gwi/api/TaskTransformationsSpec.scala
@@ -2,7 +2,7 @@ package gwi.api
 
 import akka.http.scaladsl.model.Uri
 import com.gwi.Main.ServerUri
-import com.gwi.api.{TaskDetail, TaskState}
+import com.gwi.api.{TaskDetail, TaskState, TaskTransformations}
 import com.gwi.execution.Task
 import gwi.SampleData
 import org.scalatest.flatspec.AnyFlatSpec
@@ -11,31 +11,31 @@ import org.scalatest.matchers.should
 import java.time.Instant
 import java.util.UUID
 
-class TaskDetailSpec extends AnyFlatSpec with should.Matchers with SampleData {
+class TaskTransformationsSpec extends AnyFlatSpec with should.Matchers with SampleData {
 
   val epochSecs = 1636154312
 
   "TaskDetail.getLinesRate" should "correctly calculate processed lines/s rate" in {
-    TaskDetail.getLinesRate(
+    TaskTransformations.getLinesRate(
       sampleTask.copy(startedAt = Some(Instant.ofEpochSecond(epochSecs)), endedAt = Some(Instant.ofEpochSecond(epochSecs + 2)))
     ) shouldBe 500
   }
 
   it should "properly handle division by 0" in {
-    TaskDetail.getLinesRate(
+    TaskTransformations.getLinesRate(
       sampleTask.copy(startedAt = Some(Instant.ofEpochSecond(epochSecs)), endedAt = Some(Instant.ofEpochSecond(epochSecs)))
     ) shouldBe -1
   }
 
   "TaskDetail.getMaybeResultUri" should "return some uri for Done task" in {
-    TaskDetail.getMaybeResultUri(sampleTask.copy(state = TaskState.Done)) shouldBe Some(s"$ServerUri/task/${sampleTask.id}/result")
+    TaskTransformations.getMaybeResultUri(sampleTask.copy(state = TaskState.Done)) shouldBe Some(s"$ServerUri/task/${sampleTask.id}/result")
   }
 
   it should "return empty for other states" in {
-    TaskDetail.getMaybeResultUri(sampleTask.copy(state = TaskState.Canceled)) shouldBe None
-    TaskDetail.getMaybeResultUri(sampleTask.copy(state = TaskState.Failed)) shouldBe None
-    TaskDetail.getMaybeResultUri(sampleTask.copy(state = TaskState.Running)) shouldBe None
-    TaskDetail.getMaybeResultUri(sampleTask.copy(state = TaskState.Scheduled)) shouldBe None
+    TaskTransformations.getMaybeResultUri(sampleTask.copy(state = TaskState.Canceled)) shouldBe None
+    TaskTransformations.getMaybeResultUri(sampleTask.copy(state = TaskState.Failed)) shouldBe None
+    TaskTransformations.getMaybeResultUri(sampleTask.copy(state = TaskState.Running)) shouldBe None
+    TaskTransformations.getMaybeResultUri(sampleTask.copy(state = TaskState.Scheduled)) shouldBe None
   }
 
   "TaskDetail.fromTask" should "properly task to taskDetail" in {
@@ -46,6 +46,6 @@ class TaskDetailSpec extends AnyFlatSpec with should.Matchers with SampleData {
     )
     val expectedTaskDetail =
       TaskDetail(sampleTask.id, sampleTask.linesProcessed, 1000, sampleTask.state, Some(s"$ServerUri/task/${sampleTask.id}/result"))
-    TaskDetail.fromTask(task) shouldEqual expectedTaskDetail
+    TaskTransformations.toTaskDetail(task) shouldEqual expectedTaskDetail
   }
 }
